fix: move fields initial state to server side
